TDP-6609: Backport fall 18
